FYP_Backend-main/requirements.txt
Django==5.2.8
psycopg2-binary==2.9.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
qrcode==8.0
Pillow==11.0.0
